Reformat code, delete language file

+++ template/mosaic/install.php
@@ -1,20 +1,12 @@
 <?php
 
-/*===============================================*\
-|| ############################################# ||
-|| # JAKWEB.CH                                 # ||
-|| # ----------------------------------------- # ||
-|| # Copyright 2016 JAKWEB All Rights Reserved # ||
-|| ############################################# ||
-\*===============================================*/
-
 if (!file_exists('../../config.php')) die('[index.php] config.php not found');
 require_once '../../config.php';
 
 // Check if the file is accessed only from a admin if not stop the script from running
 if (!JAK_USERID) die('You cannot access this file directly.');
 
-if(!$jakuser->jakAdminaccess($jakuser->getVar("usergroupid"))) die('You cannot access this file directly.');
+if (!$jakuser->jakAdminaccess($jakuser->getVar("usergroupid"))) die('You cannot access this file directly.');
 
 // Set successfully to zero
 $succesfully = 0;
@@ -24,37 +16,38 @@ $succesfully = 0;
 <!DOCTYPE html>
 <html lang="en">
 <head>
-	<title>Installation - Mosaic / Template</title>
-	<meta charset="utf-8">
-	<meta name="author" content="JAKWEB (http://www.jakweb.ch)" />
-	<link rel="stylesheet" href="../../css/stylesheet.css" type="text/css" media="screen" />
-	<link rel="stylesheet" href="../../css/bootstrap/bootstrap.min.css" type="text/css" media="screen" />
+  <title>Installation - Mosaic / Template</title>
+  <meta charset="utf-8">
+  <meta name="author" content="JAKWEB (http://www.jakweb.ch)"/>
+  <link rel="stylesheet" href="../../css/stylesheet.css" type="text/css" media="screen"/>
+  <link rel="stylesheet" href="../../css/bootstrap/bootstrap.min.css" type="text/css" media="screen"/>
 </head>
 <body>
 
 <div class="container">
-<div class="row">
-<div class="col-md-12">
-<h3>Installation - Mosaic / Template</h3>
+  <div class="row">
+    <div class="col-md-12">
+      <h3>Installation - Mosaic / Template</h3>
 
-<!-- Check if the plugin is already installed -->
-<?php 
+      <!-- Check if the plugin is already installed -->
+      <?php
 
-$jakdb->query('SELECT value FROM '.DB_PREFIX.'setting WHERE varname = "sitestyle_widget_mosaic"');
-if ($jakdb->affected_rows > 0) { ?>
+      $jakdb->query('SELECT value FROM ' . DB_PREFIX . 'setting WHERE varname = "sitestyle_widget_mosaic"');
+      if ($jakdb->affected_rows > 0) { ?>
 
-<div class="alert alert-info fade in">
-  Template is already installed.
-</div>
+        <div class="alert alert-info fade in">
+          Template is already installed.
+        </div>
 
-<!-- Plugin is not installed let's display the installation script -->
-<?php } else { if (isset($_POST['install'])) {
+        <!-- Plugin is not installed let's display the installation script -->
+      <?php } else {
+        if (isset($_POST['install'])) {
 
 // Delete old entries
-$jakdb->query('DELETE FROM '.DB_PREFIX.'setting WHERE product = "mosaic"');
+          $jakdb->query('DELETE FROM ' . DB_PREFIX . 'setting WHERE product = "mosaic"');
 
 // Insert tables into settings
-$jakdb->query('INSERT INTO '.DB_PREFIX.'setting (`varname`, `groupname`, `value`, `defaultvalue`, `optioncode`, `datatype`, `product`) VALUES
+          $jakdb->query('INSERT INTO ' . DB_PREFIX . 'setting (`varname`, `groupname`, `value`, `defaultvalue`, `optioncode`, `datatype`, `product`) VALUES
 ("navbarstyle_mosaic_tpl", "mosaic", 0, 0, "yesno", "boolean", "tpl_mosaic"),
 ("navbarbw_mosaic_tpl", "mosaic", "dark", "dark", "select", "free", "tpl_mosaic"),
 ("navbarcolor_mosaic_tpl", "mosaic", NULL, NULL, "input", "free", "tpl_mosaic"),
@@ -125,20 +118,22 @@ $jakdb->query('INSERT INTO '.DB_PREFIX.'setting (`varname`, `groupname`, `value`
 ("cms_tpl", "mosaic", "1", "1", "yesno", "boolean", "tpl_mosaic"),
 ("sitestyle_widget_mosaic", "mosaic", 1, 1, "yesno", "boolean", "tpl_mosaic")');
 
-$succesfully = 1;
+          $succesfully = 1;
 
-?>
-<div class="alert alert-success fade in">
-  Template successfully installed!
-</div>
-<?php } if (!$succesfully) { ?>
-<form name="company" method="post" action="install.php" enctype="multipart/form-data">
-<button type="submit" name="install" class="btn btn-primary btn-block">Install Template</button>
-</form>
-<?php } } ?>
+          ?>
+          <div class="alert alert-success fade in">
+            Template successfully installed!
+          </div>
+        <?php }
+        if (!$succesfully) { ?>
+          <form name="company" method="post" action="install.php" enctype="multipart/form-data">
+            <button type="submit" name="install" class="btn btn-primary btn-block">Install Template</button>
+          </form>
+        <?php }
+      } ?>
 
-</div>
-</div>
+    </div>
+  </div>
 
 </div>
 </body>
